requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
filelock==3.15.4
idna==3.7
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
requests-file==2.1.0
tldextract==5.1.2
urllib3==2.2.1
